Create tic tac toe
**Why** is the change needed?

My cousin wanted to build a tic tac toe game, so I thought I would make
a version so that I know how to lead him in the right direction. The
goal is not that we get the same solution, in fact that would be
disappointing, instead the goal is that I have thought through the
problem so I can give good advice.